third_party/WebKit/LayoutTests/animations/shorthand-unitless-zero.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<style>
  #test1 { animation: anim 1s 0; }
  #test2 { animation: anim 0 2s; }
</style>
<div id="test1"></div>
<div id="test2"></div>
<script>
test(function() {
  assert_equals(getComputedStyle(test1).animationIterationCount, '0');
  assert_equals(getComputedStyle(test2).animationIterationCount, '0');
  assert_equals(getComputedStyle(test1).animationDuration, '1s');
  assert_equals(getComputedStyle(test2).animationDuration, '2s');
}, 'Unitless 0 in animation shorthand should not be parsed as a delay or duration');
</script>
